Создаём 3 новых класса-менеджера: CanvasManager, ImageManager, TransformManager. Выносим методы для этих менеджеров и инитим их при создании инстанса редактора.
